For previously logged players, the new players also need to appear even if they don't move.
